third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDiffuseLightingElement-svgdom-surfaceScale-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'surfaceScale' property of the SVGFEDiffuseLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gradientElement.surfaceScale.baseVal is 0
PASS gradientElement.surfaceScale.baseVal is 1
PASS successfullyParsed is true

TEST COMPLETE
